Life Is Strange: Adjust auto bloom to work with this game
Seems there are a few differences between Viscera and Life Is Strange
despite being the same engine:

Depth buffer scaling is different, which also means the near plane
calculation is different.

Shadow shaders are not always available (much of the lighting is
pre-calculated) so use *any* available source of depth information.

Had to add the perspective divide when working out the starting offset
to check.