Viscera: Fix opacity on bloom
Adjusting the bloom position at the end meant that we missed the texture
coordinate that tells the pixel shader where to sample the depth buffer
for the opacity.

Move the adjustment to the middle of the shader, just after the view
projection matrix has been applied to adjust both, fixing the opacity on
some of the bloom.