Viscera: Fix trapdoor in 1st locked room of office (accurate UE3 multipass fog fix)
Some of the fog volumes in UE3 are a little tricky as they are rendered
in two passes that interract with each other. These can be distinguished
by trying to unadjust them - one vertex shader brings the effect to
screen depth, while the other does not.

The vertex shader that brings it to screen depth needs texcoord5 to be
adjusted, while the other needs both the position and texcoord5 to be
adjusted.

The texcoord5 adjustments can be done in either vertex or pixel shader,
but the position adjustment needs to be done in the vertex shader.

Only doing part of the adjustment will either result in an outline
around closer objects, or the fog density being incorrect from a
distance, or other problems.