Viscera: Prevent PID text depth adjustment from messing up Bitterman 500
Since we are multiplying the entire UI coordinate by a value to move it
to depth, we can prevent it from altering a given render target by
forcing that render target to mono.

Track down the render target for the Bitterman 500 and force it to mono.
Other in-game monitors seem to be ok, but if any use the same font they
might need the same treatment.

Leave the toggle key in place in case we need it, but enable the
adjustment by default.