extract_unity_shaders.py: Match hashes used by Helifax' OpenGL wrapper (TESTME)
"The way I dump the shaders:
- When the source code is loaded (before compilation) I grab it. I then
  calculate a CRC32 value over the whole information in the shader and
  write it in the filename."
    -Helifax

Try to match this scheme to (hopefully) generate the same hashes for
OpenGL shaders as Helifax' wrapper.

This has not been tested to confirm if it is accurate!

Looking at some of the OpenGL fixes, we have filenames like
"403_Vertex_40aceb.glsl". I assume the first number is an index of some
sort, which I doubt I can determine from the Unity assets, and it is
apparent that the CRC32 is *not* padded to 8 characters, so do the same.