extract_unity_shaders.py: Duple opengl shaders and insert #define FRAGMENT/VERTEX
OpenGL shaders in Unity are stored in GLSL with the vertex shader and
pixel shaders combined into one. In the .shader file this combined form
is in the vertex shader, while the pixel shader is an empty placeholder.

When they are passed to OpenGL, Unity will insert a #define
VERTEX/FRAGMENT near the top to select the appropriate type. Therefore,
we duplicate this behaviour in order to generate (we hope) identical
hashes.

We also discard the empty pixel shaders, then duplicate the vertex
shaders and change them to pixel shaders and rehash them with the
appropriate define.

All going well that should extract vertex and pixel shaders for OpenGL
Unity games and hopefully will match the hash used in Helifax' wrapper.