Far Cry 4: Fix clipping on windows & doors
The game uses a post-processing effect to shade windows and doors when
viewed from outside under certain conditions. The effect is bounded by a
2D rectangle, so at a distance the effect is clipped off in 3D.

Expand the bounding box of this effect horizontally to cover the entire
screen to remove the clipping.

We could alternatively just expand it a little, so long as the player
doesn't have the convergence set high enough for significant pop-out,
but this way it will always work.