Far Cry 4: Use resource copying to fix shadow 1 frame sync issue
The game has a bug where a shadow buffer from the previous frame is used
when anti-aliasing is disabled, or set to SMAA. In 2D this is not really
noticeable, but in 3D it has the appearance that the shadows will lift
off the ground or sink into the ground while walking.

Frame analysis was used to look for the shader that accepted the shadow
buffer as an input and drew it to the screen, and to compare it with the
output of the directional lighting shader, which identified a
discrepancy.

Use the new arbitrary resource copying support to perform our own copy
of the shadow buffer into the shader, but this time we can be sure that
we are copying the correct buffer for the current frame.