Far Cry 4: Make auto crosshair work with anti-aliasing
Getting access to the depth buffer with AA enabled turns out to be a bit
challenging.

On the one hand, if HBAO+ is enabled we can very easily get access to a
WBuffer for the depth... but this won't work with HBAO+ disabled :(

We can't use the MSAA/TXAA depth buffer directly because the crosshair
shader is SM4 and can't sample it.

We can't use 3DMigoto to resolve the anti-aliased depth buffer with
resolve_msaa because it's in a format that is not supported (in the
future we might support this via a custom shader, but not yet).

So I went back with frame analysis and found a set of shaders that the
game uses to resolve the anti-aliased depth buffer into a non-AA one and
copied the result from them. I also copy CViewportShaderParameterProvider
out of the directional lighting shaders, as it contains the values
necessary to scale the depth buffer.

Need to check if we need any other sources while underground when
directional lighting may not be enabled.