Far Cry 4: Use hud depth whenever depth buffer was not copied
Disabling the UI depth adjustment altogether whenever the depth buffer
was disabled was a bit jarring - turns that that it is enabled in some
menus some of the time and it switches around during loading.

Instead of disabling the adjustment, adjust it to the HUD depth from the
ini file. That way we still avoid misaligning anything on the full
screen map, while keeping all parts of the UI that aren't using the
depth buffer at a consistent depth.