Far Cry 4: Use integer texels instead of floats for texture coordinate filtering
Turns out the game doesn't always generate binary identical floating
point texture coordinates for each icon (at least not in the case of the
yellow triangles - so far the crosshair has been consistent), so we
can't rely on matching the texture coordinates exactly.

Instead, multiply the coordinates by the texture size and round them to
the nearest integer - that way we can be fairly certain that they should
match.