Far Cry 4: Reinvent texture coordinate filtering for simplicity
Manually determining the texture coordinates of every single icon that
we needed to adjust would be rather time consuming and error prone.
Instead, let's reinvent the technique to simplify the task to one that
can be done in under a minute!

Use the new texture injection feature in 3DMigoto to inject two textures
that correspond to the icon textures used by the game, but where I've
painted over the icons that need to be adjusted in green, and crosshairs
in blue.

In the shader we can now use the vertex buffer to find a texture
coordinate corresponding to the center of the icon, then sample the
injected textures to determine what to do.

Changing the behaviour of an icon is as simple as editing a texture :)