Far Cry 4: Use stereo2mono to improve auto HUD depth
With the new ability to copy both left and right sides of a resource
into a shader, we can improve the auto HUD depth. The original code
calculated the adjustment amount independently in each eye, which worked
surprisingly well, but was not perfect in situations where an object in
the foreground intercepted the ray in just one eye, which could cause
the crosshair to shift in one eye.

Now with the ability to check the depth buffer from both eyes
simultaneously we can change the crosshair code so that the rays from
both eyes need to intersect an object simultaneously to treat it as a
hit and move the crosshair over the object.