Binary patch Helix Mod to allow LOG.txt to be read without shutting down the game
Helix Mod is using:
  CreateFile("LOG.txt", GENERIC_ALL, FILE_SHARE_WRITE, ...);

Binary patch all possible call sites in the DLL to change it to:
  CreateFile("LOG.txt", GENERIC_READ | GENERIC_WRITE, FILE_SHARE_READ | FILE_SHARE_WRITE, ...);

Which allows the log file to be opened while Helix Mod is running.

The 32bit version is using push instructions to shovel the values onto
the stack, the 64bit version is using mov and lea instructions.