Dreamfall Chapters: Disable directional lighting as an FOV source
For some reason, using the inverse projection matrix as an FOV source is
not working in this game. Not sure if it's using the shader to render
shadows off-screen with a different FOV, or if there's a problem with
the constant copying.

Whatever, just switch to using the fallbacks as a primary source.