Dreamfall Chapters: Disable volumetric light shafts
I have some fresh ideas that might be able to fix these, but given that
I've already spend many days of effort on the problem, and even
attempting a new approach would probably take at least one or two more,
and given that the new chapter is about to be released.... forget it, at
least for now.

This disables a bunch of shaders relevant to the light shafts effect - I
could easily have gotten away with less, but this is as much to remind
me which are relevant than anything else. The list of disabled shaders
from a high level are:

Hidden_Coord
Hidden_Depth
Hidden_DepthBreaks
Hidden_Final Interpolation
Hidden_InterpolateAlongRays
Hidden_Raymarch