shadertool.py: Tweak --fix-unity-lighting-ps to find depth from matrix multiply
Shadertool was tracing the shader code forwards looking for expected
instructions to find the depth, but naturally that falls down when a
game generates a different pattern of instructions, like Dreamfall
Chapters now does. Following the instructions is somewhat more difficult
in this game, and would require relaxing some of the safety checks more
than I am comfortable with.

Instead, use the matrix multiply to determine which register and
component would contain the depth, which should be more robust against
other possible variations (this is also an approach I have proved out in
auto-fixes for Unreal).

This does result in the register being used for the depth changing,
which seems to be ok, but needs more testing to be certain that it won't
cause other issues.