extract_unity_shaders: Split --fog into --vs-fog and --ps-fog
Generating the fog variations of shaders adds a lot of additional
shaders that need to be assembled, each of which costs time to call out
to an external tool to generate.

We only really need to generate fog variations of shaders that we
actually need to fix, which is mostly going to be vertex shaders that we
run the halo autofix on - if we need fog variations of a pixel shader we
can always generate them for the specific pixel shaders.

By splitting --fog into separate options for vertex and pixel shaders we
can save quite a bit of time for very little effort.