Dreamfall Chapters: Move environment map reflections to infinity
Merely applying the Unity 5 lighting fix to the reflections moves them
to surface depth, which looks ok, but ideally we want them deeper.

Tweaking the view-space / world-space position does not look right, so
the way to adjust the reflections is to stereo correct the world space
camera position so that the reflection angles will be computed
correctly.

We use an adjustment amount for depth=0 where the camera is, then
convert it to world-space coordinates. Since the _CameraToWorld matrix
is handy we can use it if we make the adjustment in view-space
coordinates and use the view-space variation of the stereo correction
formula. We also need to use W=0 in the input coordinate to suppress any
position transformations in the matrix since we only want an adjustment
amount.