Dreamfall Chapters: Fix specular highlights that are NOT using DeferredLighting.shader
Now that we have corrected the specular highlights in Unity's various
deferred lighting shaders, it is apparent that there are still some
remaining that are drawn on the surfaces themselve. They are at surface
depth by default, which is acceptable, but since most of the highlights
are now drawn at the correct depth these look a little out of place.

These are fixed by adjusting the camera position in the pixel shader in
a similar way to the specular highlight fix in the lighting shaders. We
don't have quite the same matrices available, and while we could copy
them in let's try to make a more self-contained fix.

Instead of performing a view-space adjustment and using _CameraToWorld
to adjust the world space coordinate, perform a projection-space
adjustment and use the inverse MVP matrix (inversed via Helix Mod) to
transform it to local coordinates, then the _ObjectToWorld matrix to
transform it to world coordinates.