Dreamfall Chapters: Fix more specular highlights
Following the same pattern as the previous fix, but one of these shaders
does not have enough free outputs to pass the _ObjectToWorld matrix
through. Further - confirmed that the matrix is necessary as reflections
on the ground are broken when not using it.

Saving this commit away, but I'm going to rework this to do most of the
calculations in the vertex shader instead.