Dreamfall Chapters: Move specular highlight fix to vertex shaders
The vertex shaders are also passed the world space camera position, so
we can calculate the adjustment there without needing to pass extra
matrices to the pixel shader (we still use Helix Mod to inverse the MVP
matrix), then pass just the adjusted position to the pixel shader which
can use it to fix the specular highlights.

This allows the fix to be used in shaders that only have one free
output.