Dreamfall Chapters: Fix depth of texture drawn behind opaque windows
Some of the shops in Propast have windows that not only have a
reflection, but also a texture drawn to show the shop interior. But this
is just a texture, not a transparent glass, so it renders at surface
depth by default. These are the same shaders that have already been
adjusted for reflections + specular highlights, and using the adjusted
camera position in the vertex shader as well corrects the depth of these
textures.