Shadertool: Add --fix-unity-reflection
This option will attempt to autofix Unity object shaders with
reflections (specular and otherwise) that ordinarily render on the
surface itself.

This option must be applied to *BOTH* matching vertex and pixel shaders
for a given surface to work - it will correct the world space camera
position in the vertex shader (which by itself can be enough to fix some
effects like glass with a fake transparency texture), and pass it
through to the pixel shaders where it is used to correct reflections.

Since it may be hard to be certain that it has been applied to the
matching vertex shader (given that e.g. the texcoord may not be free, or
maybe we just forgot, or maybe the game was updated), we store a magic
value in the unused W component of the camera position. We check this
value in the pixel shader and only use it if it matches.