shadertool: Fix bug in --auto-fix-vertex-halo causing Speedtree fog fade out issue
In Mercuria in Dreamfall Chapters, I noticed as trees in the distance
reached the limit of their draw distance and were culled, the game would
fade them out, but while that happened the fog disappeared and they
became opaque for a moment.

I tracked this down to a problem in the auto halo fix, where the output
Z component was not originally written at the same time as the output X
component, but the script copied the temporary register to it anyway.

Resolve this by only searching for components that were originally
output from the relocated instruction, and ensuring that unwanted
components aren't added back in.

Also fix a second possible issue where the y and z components might have
been specified .zy instead of .yz (depending on the hash seed), which
may have rarely caused a shader not to be accepted.