Dreamfall Chapters: Add auto-HUD depth adjustment on Q key
My auto HUD code evolves once again - this is adapted from the UE3
version (since that supports adjusting based on an arbitrary coordinate)
back to work with Unity once again.

The adjustment is only applied to the icon shader and text shader, both
of which float over items in 3D. Each vertex is adjusted independently,
which can cause it to stretch over the edge of objects (similar to
compatibility mode HUDs) - some people may like it, some people may not
so add a toggle key.

There's no menu detection yet (which would be difficult as the depth
buffer is always active in this game, but I could potentially key off a
texture) - the hotkey will do for now.