shadertool: Add pixel shader centric variant of unity reflection fix
This adds a --fix-unity-reflection-ps, which performs much the same fix
as --fix-unity-reflection, but performs the adjustment in the pixel
shader instead of the vertex shader.

It still requires both vertex & pixel shaders to be adjusted since the
vertex shader is still used to pass the _Object2World matrix through to
the pixel shader (and therefore must have three free outputs), and
requires the DX9Settings.ini to be edited to inverse the MVP matrix.