Dreamfall Chapters: Change lighting fix to use yet another technique
Since we need to adjust world space coordinates for the reflection fix
and since Helix Mod only has 2 matrix copy slots and 3 constant register
copy slots let's rethink the way we fix lighting shaders in Unity so we
can use a common set of matrices between these two.

We have been applying a view-space correction in the Unity lighting
shaders, but we could just as easily apply a world-space correction
after the _CameraToWorld matrix multiplication, which we can do using
the inverted MVP and _ObjectToWorld matrices copied from elsewhere. We
can then use these same matrices for the reflection fix.

The one exception where we still need to apply a view-space correction
is in the main lighting vertex shader, but we only need to apply this
correction when drawing point or spot lights - and in those cases the
shader has valid MV and MVP matrices and can use them following the
same code as my previous Unity 5 technique to correct them, so we don't
need to use up any of Helix Mod's precious copy slots for that.

This patch adds a --fix-unity-lighting-ps-world variant to shadertool to
apply this new technique to the lighting pixel shaders, runs said script
on the lighting shaders, re-works the lighting vertex shader for the new
technique, and adds a bunch of shaders which we can use to copy the
necessary matrices.