Dreamfall Chapters: Remove directional lighting shades as Z Buffer source
It seems that the directional lighting shaders use a _ZbufferParams that
is not consistent with the rest of the game (I wonder if this might
explain the Montague's Mount quirk). I'm not entirely sure how these are
supposed to be work since they are used legitimately in the gmae (maybe
x & y are valid, while z and w are not?), but drop them as a depth
buffer source regardless.