Dreamfall Chapters: Blacklist SHADOWCASTERs as matrix sources
Narrowed down another shader that was breaking the lighting and
reflection fixes - this one was marked as a shadow caster, which
suggests that it would be rendered from the perspective of a light
source instead of the camera, which would mean it's matrices are useless
to us.

Blacklist all shadow casters as sources of the transformation matrices.