Dreamfall Chapters: Fix Seraph Kavanah sign
This is the same shader I fixed previously to work around the
inconsistent driver heuristics.

There were two problems with the workaround - firstly, it would try to
normalise the depth to equal convergence when the sign was behind the
camera would cause it to stretch over the landscape, which looked...
interesting to say the least.

Secondly, if we perform this normalisation while 3D is disabled it would
normalise it to W = convergence = 0, which would then cause it to
disappear.

Make the normalisation dependent on both of these not being the case.

The sign still appears to move slightly with the adjustment, and I'm not
sure why (floating point error perhaps?). It's not really noticeable but
would be nice if I could find a way to avoid it.