Dreamfall Chapters: Apply _WorldSpaceCameraPos fix to shaders with halo fix
Doing these separately to the last batch to make it easier to see if the
tool has done the right thing when combining the two fixes. Looks good -
declarations are being shared. Could potentially make the script share
more temp regs and the texldl, but that's a minor issue.

Removed sections to get the MVP and _Object2World matrices that were not
generated by the script - they were fine, but there would have been some
overlap since this was re-run on shaders already in the ShaderOverride
folder, and we probably don't need the extras that don't themselves
include the fix.