Dreamfall Chapters: Use FirstVertexPosReg to solve stretchy HUD
Worked out how to use Helix Mod to get access to the position of the
first vertex in a draw call so I can place all the corners of HUD icons
and entire words at a consistent depth (at least within the one draw
call).

It's not entirely ideal since it can't get access to other corners (at
least I don't think that's possible) so I can't do a minimum or average,
but it's a lot better than stretching the icons/text over the scene.