Dreamfall Chapters: Whitelist certain textures to use to adjust the HUD depth
The auto HUD depth is working pretty well, but when multiple HUD
elements are displayed at the same time they can be out of sync with
each other.

I worked out how to use FirstTexVertexPosReg in conjunction with
DefinedTexturesVS to only get the vertex position when certain textures
are matched. We can still use this from other shaders, which allows us
to move the HUD icon and text to the same depth, and means if multiple
icons are displayed at the same time (e.g. when selecting an action) the
HUD will be at a consistent depth so long as only one whitelisted icon
is displayed (or if no whitelisted icon is displayed, the position of
the most recently displayed icon will still be used).