Dreamfall Chapters: No longer blacklist IGNOREPROJECTOR shaders for matrix sources
There's at least one scene in this game (falling dreamer in chapter 1)
that only uses IGNOREPROJECTOR shaders with the matrices we need to copy
for the light and reflection fixes, therefore, we can't blacklist them.

I was never certain if this was the right keyword to blacklist anyway -
if this change breaks any other scene I'll need to investigate closer.