Dreamfall Chapters: Enable auto HUD by default and whitelist more icons
Whitelist the conversation and door icons for depth adjustment position
sources.

Unfortunately this can be displayed simultaneously with the look icon
when talking to someone, or when someone offscreen is talking, which can
misalign parts of the UI, but it's not too bad.

I'll run with the auto HUD enabled for a while and see how it feels.