Add cleanup_unity_shaders.py script to remove outdated shaders
This script will remove shaders from the fix that are no longer found
in the extracted game files. It is intended for use in games that are
updated with no way to downgrade - mostly early access titles and steam
exclusives.

This will also remove sections from the DX9Settings.ini corresponding to
any removed shaders, but it will not remove sections they refer to.

It also removes files from the installed game, and processes the
installed DX9Settings.ini as well