shadertool: Don't add command line parameters with a '*' in the vanity comment
In cygwin, the shell will expand wildcards to the individual files and I
already hide those from the vanity comment, but Windows does not do that
so I had the script glob the pattern instead. In that case it would
still add the pattern with the asterisk to the vanity comment, which I
don't really want - so filter out arguments with an '*' there too.