The Forest: Auto-adjust HUD
This only adds auto adjustments to a few icons that were originally at
screen depth.

Unfortunately several of these icons are drawn from a single draw call
which means that the first vertex position is useless (this is an area
3DMigoto has now solved), so we adjust each vertex independently of each
other. This can make the icons stretch, but it still looks pretty
reasonable.