shadertool: Fix --adjust-input=texcoord with ps_2_0 shaders
Using --adjust-input=texcoord was failing in an unexpected manner for
certain shaders. On investigation, I found that it was limited to
shaders that had been upgraded from an earlier shader model, and that
the specific problem was that the r.startswith('v') was failing.

This was due to the fact that upgrading the shader model changed r.reg
to 'v' from 't', but the string itself that the register was subclassed
from was not updated - and could not be since strings are supposed to be
immutable.

To resolve the problem we now create a new register when upgrading the
shader model instead of modifying the existing one. We then also have to
update the parent instruction which has a reference to the register in
it's args array that is separate from the syntax tree.

This commit also changes modified OpCode's to be the correct type, as I
noticed they were not as I analysed the problem - though that part
should be a non-functional change.