shadertool: Pre-check for already installed shaders with -I
Since I am now running shadertool from an extracted directory for Unity
games instead of Dumps, I'm using -I almost exclusively instead of -i,
but -I will wait until it has finished modifying a shader before it
checks if it was already installed, wasting time.

Doing a pre-check for -I is a little harder - we can't make any
assumptions about the type of shader from the path like we can for -i,
so instead we just check for the file in both possible destinations.