Stranded Deep: Apply reflection fix to Ceto shaders
This removes the use of the ocean shaders as a means to gain access to
the depth buffer - this was left over from the Unity 4 fix when the
game used it's own custom lighting shaders and I had to look elsewhere
to find the depth buffer. I don't think that is necessary any more, but
need to test.