Unity: Simplify lighting vertex shader for games using the reflection fix
Since we now only use the MV and MVP matrices in the lighting vertex
shader for non-fullscreen passes and no longer copy them from elsewhere,
we don't need to worry about the possibility of them being invalid and
we should never hit any cases where we would get a divide by zero - so
drop the safety checks to simplify the shader on the assumption that
the maths will always work out now.

We also can scrap the copy of the matrices in the temporary registers -
these were used when the matrices could have come from several different
sources, but now they only ever come from one place. We still need one
column of the model-view matrix in a temporary register due to constant
register read port limitations, if we pick MV[1] we only need the one.