shadertool: Make --ignore-other-errors ignore --adjust/--disable errors
If we are adjusting multiple inputs/outputs on several shaders in bulk
we might want to continue in the event that one of the shaders is
missing one thing we are trying to adjust, but does have at least one
other.

This makes shadertool continue if --ignore-other-errors has been
specified. It will still show the errors, and in the event that none of
the options were adjusted will still bail out on installing the shader.

Currently supported for --adjust, --unadjust, --disable, --adjust-input,
though it is not sophisticated enough to work properly if multiple of
these are specified and one fails. I could potentially change it to
behave like an autofix instead, but I don't know that I want that change
in behaviour for these.