Stranded Deep: Fix halos on water under different lighting
I had a feeling that I might need to apply different adjustments to
different pixel shaders to do the right thing, which was indeed the case
as shining a torch on the water showed a new halo, which was due to one
of these.

I've had a look at each of the vertex shaders to work out which outputs
would have been adjusted by the halo fix, and matched them to the
corresponding pixel shaders via the keywords.