The Forest: Add convergence toggle on ~
The game got rebindable keys at some point, so of course I've moved
inventory onto something sensible (tab). Move the convergence toggle to
~ so it's easy to use as well.

I could potentially put it on the same key, but the key is some
multiplayer nonsense by default, so that would be confusing to people
who try to use the devs keybindings.