shadertool: Fix Ceto reflection fix
Fixing the environment map reflections for Ceto water seems to be just a
matter of moving the sample location. The same code that worked for Life
Is Strange can work here, so refactor it and add the fix.

Notably this is not entirely accurate - I'll need to identify the render
target for the reflections and force it to stereo to do any better
though.