Stranded Deep: Add ~ as a toggle for the auto HUD
Since the game uses an unadjustable hardware cursor the crafting and
inventory menus can be difficult to use while they are at depth. Add a
key to toggle the auto crosshair off to use the menus.

Later I might use scene detection for this, but let's wait until the
game is out of early access.