Game List: Remove my username from being prominently displayed
The intent of the link to my email is just to report any issues that may
arise since I am maintaining the game list page, but I know that at
least one person has mistakenly thought I was in charge of the whole
blog because of it.

So, remove my username and just use the text "report any issues here"
instead. That should reduce the risk that someone else will have the
same misconception in future.