Shadertool: Hook up vs_2_x to use same conversion as vs_2_0
This refactors the vertex shader classes in a similar way to the pixel
shader classes, to allow the vs_2_x class to inherit from the vs_2_0
class and use the same conversion while overriding the original shader
model so the comment is correct.

The few of these I've spot checked in Black Sails look fine, but I don't
know exactly what the differences between vs_2_0 and vs_2_x are.