Far Cry 4: Fix sparks on repair tool + use more specific flame filter
The sparks from the repair tool hover in the air instead of lining up on
the vehicle. To fix this we can use the auto crosshair code to line them
up.

The sparks are drawn earlier in the frame than the current depth buffer
source, so we can't use it as is since we clear it at the start of the
frame. We could look for an alternate source or not clear it, but an
easier option is to use a second copy of the depth buffer from the
previous frame.

Remove the fire button test from one of the flames, which makes the
transition look a little better, but then it can break other flames if
the repair tool is out, so limit the distance from the camera that it
will adjust flames.