Far Cry 4: Add ability to auto adjust minimap depth
This allows all the shaders responsible for the minimap to be adjusted
from the depth buffer. Filtering is used is some of the shaders to
restrict it to HUD elements in the minimap region.

I ended up finding this was a bit distracting as the weapon tends to
wave over that part of the screen while running, so I've left it
disabled, but easily enabled via IniParams.