Unity 5 DX9: Fix distortion shader from Road Redemption
These shaders are a little hard to follow at first glance because of the
way that the instructions have been arranged, but rearranging the
instructions to put the MVP matrix multiply at the top followed by the
output writes makes it clearer (of course being careful not to break any
dependencies between instructions). Then inserting the stereo correction
follows naturally.