Far Cry 4: Blacklist icon adjustment while mission overlay is displayed
Some of the missions show an overlay when accepted with the mission name
and icon. Since the icon is the same one displayed in the world for the
mission, it is adjusted by the auto HUD code.

The mission overlay screen also has some Hindi text displayed - use
ittra-frame scene detection to disable adjusting HUD icons from the
appropriate texture while that is displayed.

This is not perfect since it takes a moment before the text is drawn
while the icon will be adjusted, but it seemed the safest option.