World of Diving: Fix lighting in Bismark level (due to fog shader)
The spot lighting in the Bismark level seemed properly aligned in terms
of shadows and so on, but the outline of the lights didn't agree in both
eyes. I determined that this was in fact due to the fog volume shader
for the water, which has a _FrustumCornersWS.

The Unity frustrum autofix did not work in this case as the shader
indexed the frustrum via a0.x, which does not work on temporary
registers. Manually fix it for now.