shadertool: Add work in progress Unity SSAO autofix
World of Diving is using an SSAO shader with a *lot* of pixel shaders.
This is a first pass attempt to fix it hoping that it would use the same
basic pattern as the other Unity SSAO shaders I've fixed. This doesn't
seem to be so simple unfortunately - this makes no visible difference,
so I'll need to investigate further.

Commit the script now, but don't enable the command line argument until
it works (it may already work on the other SSAO shaders - need to test).