Batman (likely generic UE4 pattern): Accurate shadow fix
This applies a view-space stereo correction in the computer shader that
Batman uses for shadows to move them to the correct depth.

The depth buffer and projection matrix were both identified via frame
analysis, and from there it was a fairly simple matter to locate the
coordinate constructed from the depth buffer and subtract a view-space
correction from it.