Batman: Fix density on white smoke from chimneys
Similar to the other fog shader, these need to use the uncorrected world
position for their density to agree in both eyes, so change this shader
as well. The vertex shader is used with many pixel shaders, so pass the
corrected coordinate though in texcoord5 where the original was, and the
uncorrected coordinate in a new texcoord - that way we err on the side
of fixing halos.