Batman: Fix issues caused by inserting extra texcoord at wrong spot
Inserting the extra texcoord where I did worked fine for the pair of
vertex + pixel shaders I was looking at, but broke other pixel shaders
paired with the same vertex shader thanks to fxc ignoring the texcoord
index.

Move the inserted texcoords to the end of the texcoords to that fxc just
adds a new texcoord as desired rather than rearranging the existing
ones.