Add extract_unreal_shaders.py
Since UE4 shaders are missing headers let's try to retrieve what
information we can from the game files about them. This script can
retrieve the shader name from the cooked shader cache in Arkham Knight.

Arkham Knight appears to use a custom format so this is unlikely to work
with other UE4 games as yet.

Material shaders are in a different file and are not parsed yet.

Further investigation is required to determine if it is possible to
retrieve other information like the resource names.