Infinifactory: Add additional matrix sources
shadertool will insert sections to copy matrices from any shader it
installs during a reflection fix, but won't add these sections for
other shaders. In this game, none of the autofixed shaders with the
required matrices are used in many scenes, so the matrices don't get
copied and the lights remain broken.

Add additional sources of the required matrices by script to fix the
shadows. Alternatively, the old style Unity shadow fix could have been
used as the alternative matrices that it uses are available in the
directional lighting shader.