Infinifactory: Include crosshair disable shader from bo3b
I had a quick look to see if it would be possible to use an
auto-crosshair in this game instead of disabling it, however it looks
like there are sections in the game which have no active shaders with
the required depth information (such as the tutorial, which has no
lights), and when it is available it's only in directional lighting
shaders which have proven to be unreliable in the past.

It might be possible to make it work with a bit of effort - the
directional lighting shades did have a valid _CameraDepthTexture (it was
_ZBufferParams that was messed up) so we might still be able to use them
in combination with _ZBufferParams from elsewhere. Or, we could possibly
inject the Helix Mod depth texture and fill it in from every surface
pixel shader, but experience has shown that feature is buggy.