Unity Templates: Improve documentation, add find_more_shaders.sh script
I forget just how much of the Unity fixes is still in my head - add
a README.md to the view-space variant of the Unity 5 DX9 template and
antoher README.md to the top level explaining the differences between
the two.

Also, until I improve the way that shadertool generates matrix copy
directives to separate it from the reflection fix, add a shell script
that can find additional shaders that were not auto-fixed for games
where the auto-fixed shaders were not enough.