Batman: Fix light clipping issue on small tiles
This fixes one of the two clipping issues in the Batman lighting shader.
This removes a view-space correction from the position of each tile (I
believe - it's a little hard to tell for sure) so that when their
position is compared to each of the lights it will match up and the
lights will be drawn in the correct tiles.

There is a second clipping issue that still needs to be solved - the
screen is divided up into 4x2 large tiles and the CPU has already
calculated which lights are in which of those tiles.