Batman: Fix large clipping on large 4x2 tiles in lighting shader
This creates a copy of the t10 structured buffer and merges neighboring
lists together before running the shader, which allows lights from
neighboring lists to be drawn.

Unfortunately this is a massive performance killer. While further
micro-optimisations are possible to do here, the real problem seems to
be that accessing the t10 structured buffer from so many thread groups
creates contention in the GPU. This is easy to demonstrate by limiting
the code to only do the merge on a single thread group, which restores
the original frame rate. We cannot use just a single thread group within
this shader, as the g16 buffer we are writing to is only shared within a
thread group.

I'm not sure if there is an easy way we can reduce contention within
this shader (can we declare t10 as locally coherent?), so we might need
to inject an additional shader to perform the merge from a single thread
group and write into another structured buffer to see if that makes any
improvement.