Batman: Inject a shader to perform the list merge and regain full fps
This uses the new ability to run a custom shader to solve the
performance issues created by merging the light lists in the light
shader.

The performance issue was due to reading the t10 structured buffer from
so many threads simultaneously, which created contention (unlikely to be
lock contention - much more likely to have been cache ping pong). The
writes to the g16 shared memory structured buffer had much less impact.

To do the merge in the light shader it was necessary to perform the
merge in every single thread group as the g16 shared memory structured
buffer we copied it to was only shared within each thread group. This
meant there was no easy way we could merge the lists for every thread to
see without the contention on the t10 structured buffer. On my box this
tripled the cost of the shader from 5fps to 15fps and rendered the game
unplayable.

To solve this, we inject a new custom compute shader before the lighting
compute shader to perform the merge using 8 threads of a single thread
group. This avoids the cache ping pong and the result is that the hit on
the framerate is now within the noise.