shadertool: Update patterns for Unity 5.3
Unity 5.3 has changed their asset file format (with no change to their
file headers of course :facepalm:). They no longer store register
mappings in the Unity specific headers and instead now generate
unstripped shader binaries (on account of storing the bytecode instead
of the assembly text, since it's storing it as assembly that was
stripping their headers in the first place).

This means that the CRCs of every shader will have changed yet again
(thanks Unity...), though this can probably be somewhat mitigated with
UseAlternateCRC=true in Helix Mod (which will strip headers before
calculating the CRC), though it is not clear if that works for all
shaders or only pixel shaders (needs testing).

Update the patterns in shadertool to work with either Unity or MS
style headers to enable fixing Unity 5.3 shaders.

Using MS style headers exclusively is a problem for a couple of reasons
- they don't include the shader name, and they don't include other
facts about the shader. For the purposes of shadertool, this means that
it cannot distinguish between a directional and non-directional lighting
shader, so cannot tell if the _ZBufferParams is safe to use or not, and
cannot determine if a shader is a SHADOWCASTER, so cannot tell if the
transformation matrices are safe to use or not.

To mitigate this, add some checks to make sure that Unity style headers
are attached before generating ini sections that depend on these. Until
I update the extract scripts there is no way to attach the Unity headers
back on 5.3 shaders.