Stranded Deep: Re-fix Ceto water
Ceto water has been updated and since we don't use the auto fix for
these the manual fix needed to be updated. No longer has point/spot
light variants, and it's now closer to the version used in The Forest.

With the latest update the water is broken when the camera is tilted
while side-strafing. I'm not going to spend any effort tracking that
issue down yet since the Ceto shaders are already quite problematic in
DX9 and may be updated again. Once/if I switch to DX11, or once the game
leaves Early Access I might revisit this.