Stranded Deep: Overhaul autofix.sh
Add options to easily turn various parts of the script on and off, which
is useful to focus on one part of the script that needs work.

Add a function to update the DX9Settings.ini and an option to turn it on
or off, to avoid adding duplicate sections to the ini while working on
part of the script.

Add an overwrite option for Ceto water, since we need to give it extra
special attention.