README: Add links to all games and add links to games in other branches
It looks a bit weird having links to only games on external
repositories, so add links to all the games in this repository as well.

While I'm at it, link to a few games that are in branches other than
master either due to this repo not being the master copy (Crysis 3), or
a WIP that I was unsatisfied with (Glare, Demonicon).