Introduce hlsltool.py - a DX11 shadertool
This tool aims to be a drop in replacement for shadertool.py when
working with DX11 HLSL shaders. It is lacking features compared to
shadertool, but it already implements the coveted --auto-fix-vertex-halo

It's parser is not as sophistocated as shadertool - rather than parsing
the HLSL with a tokeniser and syntax tree it merely uses regular
expressions. This is not a good general solution for parsing HLSL, but
should work ok for the subset of HLSL that 3DMigoto generates. I have no
intention of turning this into a full HLSL parser.

This tool does not support DX11 assembly shaders - eventually I would
like to extend the parser on shadertool to handle those instead.

It also strips useless x=x; instructions that MGSV generates to make the
shaders less confusing.