Batman: Fix flickering due to broken occlusion culling
The new enhancements to frame analysis logging revealed this:

1.000063 VSSetShader(pVertexShader:0x00000000C9FFC158, ppClassInstances:0x0000000000000000, NumClassInstances:0) hash=4641ff3d22219e42
1.000063 PSSetShader(pPixelShader:0x0000000000000000, ppClassInstances:0x0000000000000000, NumClassInstances:0) hash=0000000000000000
...
1.000064 Begin(pAsync:0x00000000CBDBADD0) type=query query=occlusion MiscFlags=0x0
...
1.000064 DrawIndexed(IndexCount:36, StartIndexLocation:0, BaseVertexLocation:0)
1.000065 End(pAsync:0x00000000CBDBADD0) type=query query=occlusion MiscFlags=0x0
...
2.000001 GetData(pAsync:0x00000000CBDBADD0, pData:0x000000001BE043D8, DataSize:8, GetDataFlags:1) = 0 type=query query=occlusion MiscFlags=0x0
    data: 0800000000000000

This shows that the game is querying occlusion information from the GPU
for the purposes of occlusion culling, but the occlusion culling
information is generated by running vertex shader 4641ff3d22219e42 with
no pixel shader bound.

As there is no pixel shader bound the driver will not apply the stereo
correction formula (at least not when using the batman profile), and as
such the geometry will be drawn in the 2D position when it passes the
rasteriser and may be occluded by some 3D geometry, causing the game to
skip rendering it on subsequent frames until the camera is moved or it
otherwise deems that it might be visible again.

The fix is simple once we know the correct shader - simply add the
stereo correction formula to the output position and the culling will be
calculated on the stereo positions of the geometry eliminating the
flickering.